+++ webknossos/webknossos/dataset/downsampling_utils.py
@@ -117,16 +117,9 @@ def parse_interpolation_mode(
 def linear_filter_3d(data: np.ndarray, factors: List[int], order: int) -> np.ndarray:
     factors_np = np.array(factors)
 
-    assert np.all(
-        factors_np == factors[0]
-    ), "The selected filtering strategy does not support anisotropic downsampling."
-    factor = factors[0]
-
-    ds = data.shape
-    assert not any((d % factor > 0 for d in ds))
     return zoom(
         data,
-        1 / factor,
+        1 / factors_np,
         output=data.dtype,
         # 0: nearest
         # 1: bilinear