+++ wkcuber/Changelog.md
@@ -16,7 +16,7 @@ For upgrade instructions, please check the respective *Breaking Changes* section
 ### Changed
 
 ### Fixed
-
+- Fixed that wkcuber.downsampling didn't support anisotropic downsampling for some downsampling modes like `nearest`. [#643](https://github.com/scalableminds/webknossos-libs/pull/643)
 
 ## [0.9.12](https://github.com/scalableminds/webknossos-libs/releases/tag/v0.9.12) - 2022-03-18
 [Commits](https://github.com/scalableminds/webknossos-libs/compare/v0.9.11...v0.9.12)
@@ -35,7 +35,6 @@ For upgrade instructions, please check the respective *Breaking Changes* section
 ### Fixed
 - Fixed that wkcuber.downsampling didn't support an anisotropic mag for `--from`. [#638](https://github.com/scalableminds/webknossos-libs/pull/638)
 - Fixed that wkcuber.downsampling didn't provide a meaningful error message when trying to do downsampling with an unsupported interpolation mode. [#619](https://github.com/scalableminds/webknossos-libs/pull/619)
-
 
 ## [0.9.9](https://github.com/scalableminds/webknossos-libs/releases/tag/v0.9.9) - 2022-03-03
 [Commits](https://github.com/scalableminds/webknossos-libs/compare/v0.9.8...v0.9.9)